+++ R/eager.R
@@ -4,11 +4,11 @@
 as.array.python.builtin.EagerTensor <- function(x, ...) {
   if (py_is_null_xptr(x))
     return(NULL)
-  if(x$dtype$name == "string")
+  if(as_r_value(x$dtype$name) == "string")
     array(as.character(x, ...),
           dim = if(length(dx <- dim(x))) dx else 1L)
   else
-    x$numpy()
+    as_r_value(x$numpy())
 }
 
 #' @export
@@ -107,7 +107,7 @@ as.logical.tensorflow.python.ops.variables.Variable <- as.logical.python.builtin
 
 #' @export
 as.character.python.builtin.EagerTensor <- function(x, ...) {
-  out <- x$numpy()
+  out <- as_r_value(x$numpy())
   # as.character() on python bytes dispatches to
   # reticulate:::as.character.python.builtin.bytes, which calls
   # x$decode(encoding = "utf-8", errors = "strict")
@@ -130,7 +130,7 @@ as.character.tensorflow.python.ops.variables.Variable <-
 #' @importFrom grDevices as.raster
 #' @export
 as.raster.python.builtin.EagerTensor <-
-function(x, max = if(x$dtype$is_integer) x$dtype$max else 1, ...)
+function(x, max = if(as_r_value(x$dtype$is_integer)) as_r_value(x$dtype$max) else 1, ...)
   as.raster(as.array(x), max = max, ...)
 
 #' @export